MINT-1787 Move ErrorList to src/utils
Possible problem: Using ErrorList in @chronobank/login-ui.
Now ChronoMint using aliases and ErrorList in @chronobank/login-ui
should work correctly
