+++ src/components/dashboard/DepositTokens/validate.js
@@ -4,7 +4,7 @@
  */
 
 import * as validator from '@chronobank/core/models/validator'
-import ErrorList from '@chronobank/core-dependencies/ErrorList'
+import ErrorList from 'utils/ErrorList'
 
 export default (values, props) => {
   const amount = values.get('amount')
