+++ src/components/dashboard/SendTokens/validate.js
@@ -3,7 +3,7 @@
  * Licensed under the AGPL Version 3 license.
  */
 
-import ErrorList from '@chronobank/core-dependencies/ErrorList'
+import ErrorList from 'utils/ErrorList'
 import * as validator from '@chronobank/core/models/validator'
 import tokenService from '@chronobank/core/services/TokenService'
 import { MODE_ADVANCED } from 'components/constants'
