+++ src/components/voting/PollEditForm/validate.js
@@ -3,7 +3,7 @@
  * Licensed under the AGPL Version 3 license.
  */
 
-import ErrorList from '@chronobank/core-dependencies/ErrorList'
+import ErrorList from 'utils/ErrorList'
 import BigNumber from 'bignumber.js'
 import * as validator from '@chronobank/core/models/validator'
